+++ src/lib/pages/validator-details/components/bonded-token-changes/VotingPowerChart.tsx
@@ -1,9 +1,8 @@
-import { Box, Button, Flex, Heading, Text } from "@chakra-ui/react";
+import { Box, Button, Flex } from "@chakra-ui/react";
 import type { BigSource } from "big.js";
 import type { ScriptableContext, TooltipModel } from "chart.js";
 
-import { TabIndex } from "../../types";
-import { useInternalNavigate } from "lib/app-provider";
+import { useMobile } from "lib/app-provider";
 import { LineChart } from "lib/components/chart/LineChart";
 import { CustomIcon } from "lib/components/icon";
 import { Loading } from "lib/components/Loading";
@@ -16,20 +15,23 @@ import {
   getTokenLabel,
 } from "lib/utils";
 
+import { VotingPowerChartDetails } from "./VotingPowerChartDetails";
+
 interface VotingPowerChartProps {
   validatorAddress: ValidatorAddr;
   singleStakingDenom: Option<string>;
   assetInfos: Option<AssetInfos>;
-  isOverview?: boolean;
+  onViewMore?: () => void;
 }
 
 export const VotingPowerChart = ({
   validatorAddress,
   singleStakingDenom,
   assetInfos,
-  isOverview,
+  onViewMore,
 }: VotingPowerChartProps) => {
-  const navigate = useInternalNavigate();
+  const isMobile = useMobile();
+  const isMobileOverview = isMobile && !!onViewMore;
 
   const { data: historicalPowers, isLoading } =
     useValidatorHistoricalPowers(validatorAddress);
@@ -37,11 +39,35 @@ export const VotingPowerChart = ({
   if (isLoading) return <Loading />;
   if (!historicalPowers) return <ErrorFetching dataName="historical powers" />;
 
-  const labels = historicalPowers?.items.map((item) =>
+  const assetInfo = singleStakingDenom
+    ? assetInfos?.[singleStakingDenom]
+    : undefined;
+
+  if (isMobileOverview)
+    return (
+      <Flex
+        backgroundColor="gray.900"
+        p={4}
+        rounded={8}
+        w="100%"
+        justifyContent="space-between"
+        alignItems="center"
+        onClick={onViewMore}
+      >
+        <VotingPowerChartDetails
+          historicalPowers={historicalPowers}
+          singleStakingDenom={singleStakingDenom}
+          assetInfo={assetInfo}
+        />
+        <CustomIcon boxSize={6} m={0} name="chevron-right" color="gray.600" />
+      </Flex>
+    );
+
+  const labels = historicalPowers.items.map((item) =>
     formatHHmm(item.hourRoundedTimestamp as Date)
   );
 
-  const dateLabels = labels?.map((label) => {
+  const dateLabels = labels.map((label) => {
     const [hours, minutes] = label.split(":");
     const date = new Date();
     date.setHours(parseInt(hours, 10));
@@ -52,8 +78,7 @@ export const VotingPowerChart = ({
   });
 
   const dataset = {
-    data:
-      historicalPowers?.items.map((item) => item.votingPower.toNumber()) ?? [],
+    data: historicalPowers.items.map((item) => item.votingPower.toNumber()),
     borderColor: "#D8BEFC",
     backgroundColor: (context: ScriptableContext<"line">) => {
       const { ctx } = context.chart;
@@ -69,32 +94,19 @@ export const VotingPowerChart = ({
     pointHoverBorderColor: "#D8BEFC",
   };
 
-  const assetInfo = singleStakingDenom
-    ? assetInfos?.[singleStakingDenom]
-    : undefined;
-
-  const handleFormatValue = (value: string | number, isSuffix = true) =>
-    formatUTokenWithPrecision(
-      value as U<Token<BigSource>>,
-      assetInfo?.precision ?? 0,
-      isSuffix,
-      2
-    );
-
-  const isDatasetContainsData = dataset && dataset.data.length > 0;
-
   const currency = singleStakingDenom
     ? `${getTokenLabel(singleStakingDenom, assetInfo?.symbol)}`
     : "";
-  const currentPrice = isDatasetContainsData
-    ? handleFormatValue(dataset.data[dataset.data.length - 1])
-    : "";
-  const diffInLast24Hr = isDatasetContainsData
-    ? dataset.data[dataset.data.length - 1] - dataset.data[0]
-    : 0;
 
   const customizeTooltip = (tooltip: TooltipModel<"line">) => {
     const { raw, dataIndex } = tooltip.dataPoints[0];
+
+    const formattedAmount = formatUTokenWithPrecision(
+      raw as U<Token<BigSource>>,
+      assetInfo?.precision ?? 0,
+      false,
+      2
+    );
 
     return `
       <div style="padding: 8px 12px;">
@@ -102,7 +114,7 @@ export const VotingPowerChart = ({
           <h1 style="font-size: 12px; color: #ADADC2;">${
             singleStakingDenom ? "Bonded Token" : "Voting Powers"
           }</h1>
-          <p style="font-size: 16px; color: #F7F2FE; white-space: nowrap;">${handleFormatValue(raw as number, false)} ${currency}</p>
+          <p style="font-size: 16px; color: #F7F2FE; white-space: nowrap;">${formattedAmount} ${currency}</p>
         </div>
         <hr style="margin-top: 8px; color: #68688A;"/>
         <p style="margin-top: 8px; font-size: 12px; color: #F7F2FE; white-space: nowrap;">${dateLabels[dataIndex]}</p>
@@ -123,44 +135,18 @@ export const VotingPowerChart = ({
       w="100%"
     >
       <Flex gap={6} direction="column" w={250} minW={250}>
-        <Flex gap={2} direction="column">
-          <Heading variant="h6">
-            {singleStakingDenom
-              ? "Current Bonded Token"
-              : "Current Voting Powers"}
-          </Heading>
-          <Heading variant="h5" fontWeight={600}>
-            {currentPrice} {currency}
-          </Heading>
-          <Text variant="body1">
-            <Text
-              as="span"
-              fontWeight={700}
-              color={diffInLast24Hr >= 0 ? "success.main" : "error.main"}
-            >
-              {diffInLast24Hr >= 0
-                ? `+${handleFormatValue(diffInLast24Hr)}`
-                : `-${handleFormatValue(-diffInLast24Hr)}`}
-            </Text>{" "}
-            {currency} in last 24 hr
-          </Text>
-        </Flex>
-        {isOverview && (
+        <VotingPowerChartDetails
+          historicalPowers={historicalPowers}
+          singleStakingDenom={singleStakingDenom}
+          assetInfo={assetInfo}
+        />
+        {onViewMore && (
           <Button
             variant="ghost-secondary"
             p="unset"
             size="md"
             pl={2}
-            onClick={() =>
-              navigate({
-                pathname: "/validators/[validatorAddress]/[tab]",
-                query: {
-                  validatorAddress,
-                  tab: TabIndex.BondedTokenChanges,
-                },
-                options: { shallow: true },
-              })
-            }
+            onClick={onViewMore}
           >
             See all related transactions
             <CustomIcon name="chevron-right" boxSize={3} />
@@ -172,7 +158,14 @@ export const VotingPowerChart = ({
           labels={labels}
           dataset={dataset}
           customizeTooltip={customizeTooltip}
-          customizeYAxisTicks={handleFormatValue}
+          customizeYAxisTicks={(value) =>
+            formatUTokenWithPrecision(
+              value as U<Token<BigSource>>,
+              assetInfo?.precision ?? 0,
+              false,
+              2
+            )
+          }
         />
       </Box>
     </Flex>